# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dự án tốt nghiệp",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/auth/signup": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Đăng ký",
                "operationId": "af6956352ace90b33a7bc69f77eb585e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "confirm_password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "default": "your name"
                                    },
                                    "email": {
                                        "type": "string",
                                        "default": "your_email@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "default": "Abc1234"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "default": "Abc1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "validate"
                    },
                    "500": {
                        "description": "server"
                    }
                }
            }
        },
        "/auth/verifyOTP": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Verify OTP",
                "description": "This endpoint verifies an OTP sent to the user's email.",
                "operationId": "9ca8272abe2e190aea595bedab029ebf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "OTP",
                                    "email"
                                ],
                                "properties": {
                                    "OTP": {
                                        "description": "The OTP to verify.",
                                        "type": "string",
                                        "format": "digits",
                                        "example": "1234"
                                    },
                                    "email": {
                                        "description": "The email address associated with the OTP.",
                                        "type": "string",
                                        "format": "email",
                                        "example": "your_email@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP verified successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "OTP verified successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid OTP or email."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "login",
                "description": "This endpoint login",
                "operationId": "9da2cf5088492a12575eabde8425e309",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "The user's email.",
                                        "type": "string",
                                        "format": "email",
                                        "example": "your_email@example.com"
                                    },
                                    "password": {
                                        "description": "The user's password.",
                                        "type": "string",
                                        "format": "password",
                                        "example": "Abc1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "login successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "login successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid email."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/refreshToken": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "auth Resource",
                "description": "refreshToken",
                "operationId": "abd35f1b78003136dde1a06a96b39a07",
                "responses": {
                    "200": {
                        "description": "Successful retrieval of protected resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "A protected resource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "auth Resource",
                "description": "Logout",
                "operationId": "641d632667d8cd0019e6d3c7d802657a",
                "responses": {
                    "200": {
                        "description": "Successful retrieval of protected resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "A protected resource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/user/profile": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Protected Resource",
                "description": "User",
                "operationId": "8fa93d3cc0f8379a388c26a467c13cbc",
                "responses": {
                    "200": {
                        "description": "Successful retrieval of protected resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "A protected resource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Enter JWT Bearer token **_only_**",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}